+++ libsolidity/analysis/TypeChecker.cpp
@@ -1543,6 +1543,8 @@ void TypeChecker::endVisit(NewExpression const& _newExpression)
 
 		if (!contract)
 			m_errorReporter.fatalTypeError(_newExpression.location(), "Identifier is not a contract.");
+		if (contract->contractKind() != ContractDefinition::ContractKind::Contract)
+			    m_errorReporter.fatalTypeError(_newExpression.location(), "Cannot instantiate an interface or a library.");
 		if (!contract->annotation().unimplementedFunctions.empty())
 			m_errorReporter.typeError(
 				_newExpression.location(),