+++ Changelog.md
@@ -8,6 +8,7 @@ Features:
 
 Bugfixes:
  * Parser: Enforce commas between array and tuple elements.
+ * Type Checker: Reject the creation of interface and library contracts using the ``new`` statement.
 
 ### 0.4.15 (2017-08-08)